Move assertion utils to separate file due to circular dependency with process_client
