+++ src/layman/geoserver_proxy_test.py
@@ -1,5 +1,6 @@
 import time
-from test import process_client as client_util, geoserver_client, util as test_util
+
+from test import process_client as client_util, geoserver_client, util as test_util, assert_util
 from test.process_client import get_authz_headers
 from test.data import wfs as data_wfs, SMALL_LAYER_BBOX
 import requests
@@ -7,8 +8,7 @@ from owslib.feature.schema import get_schema as get_wfs_schema
 import pytest
 
 from geoserver.util import get_layer_thumbnail, get_square_bbox
-from layman import app, settings, util as layman_util
-from layman.common import bbox as bbox_util
+from layman import app, settings
 from layman.layer import db, util as layer_util
 from layman.layer.filesystem import thumbnail
 from layman.layer.geoserver import wfs as geoserver_wfs
@@ -369,28 +369,6 @@ def test_missing_attribute_authz():
     client_util.delete_workspace_layer(username, layername1, headers=headers1)
 
 
-def assert_all_sources_bbox(workspace, layer, expected_bbox):
-    with app.app_context():
-        bbox = tuple(layman_util.get_publication_info(workspace, client_util.LAYER_TYPE, layer,
-                                                      context={'key': ['bounding_box']})['bounding_box'])
-    test_util.assert_same_bboxes(expected_bbox, bbox, 0)
-    test_util.assert_wfs_bbox(workspace, layer, expected_bbox)
-    test_util.assert_wms_bbox(workspace, layer, expected_bbox)
-
-    with app.app_context():
-        expected_bbox_4326 = bbox_util.transform(expected_bbox, 3857, 4326, )
-    md_comparison = client_util.get_workspace_layer_metadata_comparison(workspace, layer)
-    csw_prefix = settings.CSW_PROXY_URL
-    csw_src_key = client_util.get_source_key_from_metadata_comparison(md_comparison, csw_prefix)
-    assert csw_src_key is not None
-    prop_key = 'extent'
-    md_props = md_comparison['metadata_properties']
-    assert md_props[prop_key]['equal'] is True
-    assert md_props[prop_key]['equal_or_null'] is True
-    csw_bbox_4326 = tuple(md_props[prop_key]['values'][csw_src_key])
-    test_util.assert_same_bboxes(expected_bbox_4326, csw_bbox_4326, 0.001)
-
-
 @pytest.mark.parametrize('style_file, thumbnail_style_postfix', [
     (None, '_sld'),
     ('sample/style/small_layer.qml', '_qml'),
@@ -402,7 +380,7 @@ def test_wfs_bbox(style_file, thumbnail_style_postfix):
 
     client_util.publish_workspace_layer(workspace, layer, style_file=style_file, )
 
-    assert_all_sources_bbox(workspace, layer, SMALL_LAYER_BBOX)
+    assert_util.assert_all_sources_bbox(workspace, layer, SMALL_LAYER_BBOX)
 
     rest_url = f"http://{settings.LAYMAN_SERVER_NAME}/geoserver/{workspace}/wfs?request=Transaction"
     headers = {
@@ -427,7 +405,7 @@ def test_wfs_bbox(style_file, thumbnail_style_postfix):
         # until there is way to check end of asynchronous task after WFS-T
         time.sleep(5)
 
-        assert_all_sources_bbox(workspace, layer, exp_bbox)
+        assert_util.assert_all_sources_bbox(workspace, layer, exp_bbox)
 
         expected_thumbnail_path = f'/code/sample/style/{layer}{thumbnail_style_postfix}{thumbnail_bbox_postfix}.png'
         with app.app_context():
